chore: hoist trait test utilities
